archived/langchain-docs-benchmarking/pyproject.toml
[tool.poetry]
name = "langservehub-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langsmith = ">=0.0.64,<0.1.0"
sse-starlette = "^1.6.5"
tomli-w = "^1.0.0"
uvicorn = "^0.23.2"
fastapi = "^0.104"
langserve = ">=0.0.16"
chat-langchain = {path = "packages/chat-langchain", develop = true}
langchain-docs-retriever = {path = "packages/langchain-docs-retriever", develop = true}
anthropic-iterative-search = {path = "packages/anthropic-iterative-search", develop = true}
oai-assistant = {path = "packages/oai-assistant", develop = true}
openai-functions-agent = {path = "packages/openai-functions-agent", develop = true}

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.23.2"
pygithub = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
